Remove try/except import blocks for external packages
Replace every import-protecting try/except across the repo with a
direct top-level import, and add the now-required packages to
environment_new.yml.

evoaug (3 files): use the user's evoaug2==2.0.4 PyPI package, which
installs as Python module `evoaug` so the existing
`from evoaug.augment import ...` / `from evoaug.evoaug import RobustLoader`
lines work unchanged. Drop EvoAug_AVAILABLE / EVOAUG_AVAILABLE flags and
all guard branches that fall back to standard dataloaders or empty
augmentation lists when the flag is False.

flash_attn.layers.rotary fallback in model/rotary.py: deleted. The pinned
flash-attn==2.7.4.post1 ships layers.rotary, so the torchscript fallback
(_apply_rotary_pos_emb_torchscript + rotate_half) was unreachable.

pytorch_lightning fallback in scripts/shuffled_evaluations.py: deleted.
pytorch-lightning is required; the dummy `pl` type object the fallback
created would silently break any module that hit it.

fsspec/pyfaidx in model_zoo/ccre/traitgym_benchmark.py: imports hoisted
to the top of the file. The runtime try/except around the actual
Fasta(fsspec.open(...)) call stays since it catches S3 I/O errors, not
import errors.

wandb in scripts/sample.py: dropped the ImportError fallback. wandb is
required in env. The use_wandb flag still gates whether anything gets
logged.

Also extract utils/dinuc_shuffle.py: previously dinuc_shuffle was
defined inline in scripts/shuffled_evaluations.py, but
scripts/deepstarr_attributions_template.py was importing it from
utils.dinuc_shuffle (a module that didn't exist). Moved the function to
a real utility module so both consumers can import it cleanly.

environment_new.yml: added evoaug2==2.0.4, fsspec==2023.10.0,
pyfaidx==0.8.1.1, s3fs==2023.10.0.
