+++ model_zoo/lentimpra/iterative_aug.py
@@ -38,16 +38,11 @@ from scipy import stats
 from pytorch_lightning import loggers as pl_loggers
 from pytorch_lightning.callbacks import ModelCheckpoint, EarlyStopping
 
-# EvoAug imports
-try:
-    from evoaug.augment import (
-        RandomDeletion, RandomRC, RandomInsertion,
-        RandomTranslocation, RandomMutation, RandomNoise
-    )
-    from evoaug.evoaug import RobustLoader
-    EVOAUG_AVAILABLE = True
-except Exception:
-    EVOAUG_AVAILABLE = False
+from evoaug.augment import (
+    RandomDeletion, RandomRC, RandomInsertion,
+    RandomTranslocation, RandomMutation, RandomNoise,
+)
+from evoaug.evoaug import RobustLoader
 
 # Add project root to Python path
 project_root = Path(__file__).parent.parent.parent
@@ -365,8 +360,8 @@ def training_with_PL(dataset_path: str,
         train_dataset = TensorDataset(X_train, Y_train)
         val_dataset = TensorDataset(X_val, Y_val)
 
-        # Use EvoAug RobustLoader for training if requested and available
-        if use_augmentation and EVOAUG_AVAILABLE:
+        # Use EvoAug RobustLoader for training if requested
+        if use_augmentation:
             # Augmentation list matching DeepSTARR (lines 446-451)
             augment_list = [
                 RandomDeletion(delete_min=0, delete_max=20),
@@ -397,9 +392,6 @@ def training_with_PL(dataset_path: str,
                 pin_memory=True
             )
         else:
-            if use_augmentation and not EVOAUG_AVAILABLE:
-                print("Warning: EvoAug requested but not available. Falling back to standard dataloaders.")
-
             # Standard DataLoaders
             train_dataloader = DataLoader(
                 train_dataset,
@@ -457,7 +449,7 @@ def training_with_PL(dataset_path: str,
         trainer.fit(model, train_dataloaders=train_dataloader, val_dataloaders=val_dataloader)
 
         # Finetune on the original data without augmentations (if augmentation was used)
-        if use_augmentation and EVOAUG_AVAILABLE:
+        if use_augmentation:
             # Build non-augmented dataloaders for finetuning
             finetune_train_loader = DataLoader(
                 TensorDataset(X_train, Y_train),
